KokkosBatched - performance test host gemm
