--- perf_test/batched/host/KokkosKernels_config.h
@@ -1,1 +1,0 @@
-// Dummy
