+++ src/batched/KokkosBatched_Gemv_Team_Internal.hpp
@@ -48,9 +48,6 @@ namespace KokkosBatched {
            const ValueType *__restrict__ x, const int xs0,
            const ScalarType beta,
            /**/  ValueType *__restrict__ y, const int ys0) {
-      static_assert(is_same_mag_type<ScalarType,ValueType>::value && !is_vector<ScalarType>::value,
-                    "TeamGemvInternal:: not valid template types");
-
       const ScalarType one(1.0), zero(0.0);
 
       // y = beta y + alpha A x
@@ -93,8 +90,6 @@ namespace KokkosBatched {
            const ValueType *__restrict__ x, const int xs0,
            const ScalarType beta,
            /**/  ValueType *__restrict__ y, const int ys0) {
-      static_assert(is_same_mag_type<ScalarType,ValueType>::value && !is_vector<ScalarType>::value,
-                    "TeamGemvInternal:: not valid template types");
 
       const ScalarType one(1.0), zero(0.0);
