+++ src/batched/KokkosBatched_Set_Internal.hpp
@@ -20,8 +20,6 @@ namespace KokkosBatched {
       invoke(const int m, 
              const ScalarType alpha, 
              /* */ ValueType *__restrict__ A, const int as0) {
-        static_assert(is_same_mag_type<ScalarType,ValueType>::value && !is_vector<ScalarType>::value,
-                      "SerialSetInternal:: not valid template types");
 
 #if defined(KOKKOS_ENABLE_PRAGMA_UNROLL)
 #pragma unroll
@@ -63,8 +61,6 @@ namespace KokkosBatched {
              const int m, 
              const ScalarType alpha, 
              /* */ ValueType *__restrict__ A, const int as0) {
-        static_assert(is_same_mag_type<ScalarType,ValueType>::value && !is_vector<ScalarType>::value,
-                      "SerialSetInternal:: not valid template types");
         
         Kokkos::parallel_for
           (Kokkos::TeamThreadRange(member,0,m),
