+++ src/batched/KokkosBatched_Trsv_Serial_Internal.hpp
@@ -48,8 +48,6 @@ namespace KokkosBatched {
            const ScalarType alpha,
            const ValueType *__restrict__ A, const int as0, const int as1,
            /**/  ValueType *__restrict__ b, const int bs0) {
-      static_assert(is_same_mag_type<ScalarType,ValueType>::value && !is_vector<ScalarType>::value,
-                    "SerialTrsvInternal:: not valid template types");
 
       const ScalarType one(1.0), zero(0.0);
 
@@ -89,8 +87,6 @@ namespace KokkosBatched {
            const ScalarType alpha,
            const ValueType *__restrict__ A, const int as0, const int as1,
            /**/  ValueType *__restrict__ b, const int bs0) {
-      static_assert(is_same_mag_type<ScalarType,ValueType>::value && !is_vector<ScalarType>::value,
-                    "SerialTrsvInternal:: not valid template types");
 
       const ScalarType one(1.0), zero(0.0), minus_one(-1.0);
 
@@ -160,8 +156,6 @@ namespace KokkosBatched {
            const ScalarType alpha,
            const ValueType *__restrict__ A, const int as0, const int as1,
            /**/  ValueType *__restrict__ b, const int bs0) {
-      static_assert(is_same_mag_type<ScalarType,ValueType>::value && !is_vector<ScalarType>::value,
-                    "SerialTrsvInternal:: not valid template types");
 
       const ScalarType one(1.0), zero(0.0);
 
@@ -198,8 +192,6 @@ namespace KokkosBatched {
            const ScalarType alpha,
            const ValueType *__restrict__ A, const int as0, const int as1,
            /**/  ValueType *__restrict__ b, const int bs0) {
-      static_assert(is_same_mag_type<ScalarType,ValueType>::value && !is_vector<ScalarType>::value,
-                    "SerialTrsvInternal:: not valid template types");
 
       const ScalarType one(1.0), zero(0.0), minus_one(-1.0);
